fixing defect.

+++ 07-reading-text-files/FileWordCounter.java
@@ -3,7 +3,7 @@ import java.util.Map;
 
 public class FileWordCounter {
     
-    public Map<String, Integer> countWords(String fileName) {
+    public Map<String, Long> countWords(String fileName) {
         return Collections.emptyMap();
     }
